+++ src/main/java/lv/lumii/obis/schema/services/dto/SchemaClassNodeInfo.java
@@ -10,7 +10,7 @@ import java.util.List;
 public class SchemaClassNodeInfo {
 	
 	private String className;
-	private Integer instanceCount;
+	private Long instanceCount;
 	private List<String> neighbors;
 	
 	public List<String> getNeighbors() {
